Split tests according to divided rules
